+++ CHANGELOG.md
@@ -5,10 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### [0.0.22]
 
 #### Fixed
+- Fixed date parsing for RouteType.Blog.
+- Fix NPE in prepareFragments (#1).
+- Correctly setup Flexmark for processing Markdown tables.
+- Workaround for \u2028 characters appearing in output generated from Markdown tables.
 
 #### Changed
-- Updated flexmark to 0.32.20.
-- Fixed date parsing for RouteType.Blog.
+- Updated flexmark to 0.32.56.
 
 #### Added